feat(PostRepositoryCustom): 커스텀 리포지토리 작성
bookId로 post 리스트를 찾는 search 메서드 작성

+++ src/main/java/com/Hajuuu/page/repository/PostRepositoryCustom.java
@@ -5,5 +5,5 @@ import java.util.List;
 
 public interface PostRepositoryCustom {
 
-    List<PostFormDTO> search(Long bookId);
+    List<PostFormDTO> search(int bookId);
 }
